+++ data/people.json
@@ -1,0 +1,1 @@
+[{"type":"Student","name":"dsd","age":21,"classroom":true,"permission":"11"}]
