+++ data/rentals.json
@@ -1,0 +1,1 @@
+[{"book_title":"dsad","person_id":591,"date":"2020-02-21"}]
